[#151780617] Add initializer
