+++ lib/generators/templates/initializers/simple_admin.rb.erb
@@ -1,0 +1,11 @@
+SimpleAdmin::Menu.configure do
+  menu_item do
+    title { 'Test' }
+    route { 'Test' }
+  end
+
+  menu_item do
+    title { 'Another' }
+    route { 'Another' }
+  end
+end
